==> SCRIPTS/0.1 ANALYSIS MASTER SCRIPT.R <==

# Master script for analysis of data for chapter 3

# Prep workspace

source("SCRIPTS/1.0 LOAD PACKAGES.R")# Loads required packages
source("SCRIPTS/1.1 PREP DATA.R")# Extracts data from databases and saves environment


# Plots & tables of WQ data

source("SCRIPTS/2.0 PLOTS & TABLES OF WQ DATA.R")
