==> SCRIPTS/1.0 LOAD PACKAGES.R <==

#prepws

library(tidyverse)
library(lubridate)
library(caret)
library(purrr)
library(broom)
library(tidyr)
library(GGally)
library(rowr)
library(dbplyr)
library(RCurl)
library(repmis)
library(RcppRoll)
library(doParallel)
library(pracma)
library(zoo)
library(stargazer)
library(MASS) 
library(forecast)
library(tseries)
library(sqldf)
library(mtsdi)
library(missForest)
library(mgcv)
library(voxel)
library(lmtest)
library(itsadug)
library(cowplot)
library(egg)


registerDoParallel(cores = detectCores(all.tests = FALSE, logical = TRUE))

citation("tidyverse")
citation("lubridate")
citation("caret")
citation("purrr")
citation("broom")
citation("tidyr")
citation("GGally")
citation("rowr")
citation("dbplyr")
citation("RCurl")
citation("repmis")
citation("RcppRoll")
citation("doParallel")
citation("pracma")
citation("zoo")
citation("stargazer")
citation("MASS")
citation("base")
citation("superpc")
citation("stats")
